Move `QueryValidation` component to searchbar directory.
