+++ graylog2-web-interface/src/views/components/searchbar/QueryValidation.tsx
